NewsPaper/requirements.txt
asgiref==3.8.1
certifi==2025.1.31
cffi==1.17.1
charset-normalizer==3.4.1
check-swear==0.1.4
click==8.1.7
colorama==0.4.6
cryptography==44.0.0
Django==5.1.3
django-allauth==65.4.1
django-filter==24.3
idna==3.10
joblib==1.3.2
nltk==3.9.1
numpy==1.26.4
psycopg2==2.9.10
pycparser==2.22
PyJWT==2.10.1
python-dotenv==1.0.1
regex==2024.11.6
requests==2.32.3
scikit-learn==1.4.0
scipy==1.14.1
six==1.17.0
sqlparse==0.5.1
threadpoolctl==3.5.0
tqdm==4.67.1
tzdata==2024.2
urllib3==2.3.0